feat(sentry): rename sentry util popup -> react-error-boundary
